% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_boot.R
\name{BC_boot}
\alias{BC_boot}
\title{BC_boot}
\usage{
BC_boot(x, nboot)
}
\arguments{
\item{x}{Generic name for a vector of data to bootstrap sample BCs}

\item{alpha}{significnace level}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   bootstrapped p-value
}\if{html}{\out{</div>}}
}
\description{
BC_boot
}
