% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{calc_power}
\alias{calc_power}
\title{calculate power}
\usage{
calc_power(n, p1, p2, shift, sel, CI, alpha = 0.05, nsim = 100)
}
\arguments{
\item{n}{number of something}

\item{p1}{pr(event) in subgroup 1}

\item{p2}{pr(event) in subgroup 2}

\item{shift}{scaled distance between means}

\item{sel}{column to select}

\item{CI}{generate a CI (?)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (100)}
}
\description{
calculate power
}
