% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_params}
\alias{convert_params}
\title{convert_params}
\usage{
convert_params(dist, params)
}
\arguments{
\item{dist}{The distribution family}

\item{params}{2 level list with mu and sd for both modes or samples}

\item{n}{number of samples. This should have two elements}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    params.transform
}\if{html}{\out{</div>}}
}
\description{
This function takes the list of mean and SD parameters, as well as the distribution ution family,
log-normal, gaussian, gamma, beta, or Weibull; and converts the mean and SD into the respective distribution parameters.
}
