% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_2_clust_copula.R
\name{est_pow_2_clust_copula}
\alias{est_pow_2_clust_copula}
\title{est_pow_2_clust_copula}
\usage{
est_pow_2_clust_copula(
  n,
  alpha = 0.05,
  nsim = 20,
  dist = c("norm", "beta", "weibull", "gamma", "lnorm"),
  params,
  rho = c(0, 0),
  tests
)
}
\arguments{
\item{n}{number of somethings c(cluster 1, cluster 2)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}

\item{dist}{generating distribution}

\item{params}{parameters for the generating distribution}

\item{tests}{names of tests to run}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a power estimate
}\if{html}{\out{</div>}}
}
\description{
est_pow_2_clust_copula
}
