% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_2_mixtures.R
\name{est_pow_2_mixtures}
\alias{est_pow_2_mixtures}
\title{est_pow_2_mixtures}
\usage{
est_pow_2_mixtures(
  n_group1,
  n_group2,
  alpha = 0.05,
  nsim = 20,
  dist = c("norm", "beta", "weibull", "gamma", "lnorm"),
  params,
  tests
)
}
\arguments{
\item{n_group1}{number of samples in c(group 1 mode 1 , group 1 mode 2)}

\item{n_group2}{number of samples in c(group 2 mode 1, group 2 mode 2)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}

\item{dist}{generating distributions}

\item{params}{parameters for the generating distributions}

\item{tests}{names of tests to run between the two groups}

\item{modes}{modes of the two groups}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a power estimate
}\if{html}{\out{</div>}}
}
\description{
est_pow_2_mixtures
}
