% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_bimodal.R
\name{est_pow_bimodal}
\alias{est_pow_bimodal}
\title{est_pow_bimodal}
\usage{
est_pow_bimodal(
  n,
  alpha = 0.05,
  nsim = 20,
  dist = c("norm", "beta", "weibull", "gamma", "lnorm"),
  params,
  tests
)
}
\arguments{
\item{n}{number of samples c(mode 1, mode 2)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}

\item{dist}{generating distribution}

\item{params}{parameters for the generating distribution}

\item{tests}{names of tests to run}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a power estimate
}\if{html}{\out{</div>}}
}
\description{
est_pow_bimodal
}
