% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_bin.R
\name{est_pow_bin}
\alias{est_pow_bin}
\title{est_pow_bin}
\usage{
est_pow_bin(n, p, alpha = 0.05, nsim = 20)
}
\arguments{
\item{n}{number of \link{somethings}}

\item{p}{pr(event)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}
}
\value{
\preformatted{   a power estimate
}
}
\description{
est_pow_bin
}
