% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_het.R
\name{est_pow_het}
\alias{est_pow_het}
\title{est_pow_het}
\usage{
est_pow_het(n, p, alpha, x, nsim)
}
\arguments{
\item{n}{number of \link{somethings}}

\item{p}{pr(event)}

\item{alpha}{default significance level (0.05)}

\item{x}{I have no idea what this does}

\item{nsim}{number of simulations (20)}
}
\value{
\preformatted{   a power estimate
}
}
\description{
est_pow_het
}
