% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dist.R
\name{find_dist}
\alias{find_dist}
\title{find_dist}
\usage{
find_dist(n1, n2, alpha, nsim)
}
\arguments{
\item{n1}{number of \link{somethings} in group 1}

\item{n2}{number of \link{somethings} in group 2}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a data.frame
}\if{html}{\out{</div>}}
}
\description{
find_dist
}
