% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniDiff.R
\name{giniDiff}
\alias{giniDiff}
\title{giniDiff}
\usage{
giniDiff(y, X)
}
\arguments{
\item{y}{Generic name for the outcome measures}

\item{X}{Generic name for the 2-level grouping variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Absolute value of the difference between group 1 and group 2 Gini's mean difference
}\if{html}{\out{</div>}}
}
\description{
giniDiff
}
