% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initz2.R
\name{initz2}
\alias{initz2}
\title{Our own initz with fuzzy c means if kmeans or hclust fails in mixR::initz}
\usage{
initz2(x, ncomp, init.method = c("kmeans", "hclust", "fuzzy"), minprop = 0.05)
}
\arguments{
\item{x}{the data to fit}

\item{ncomp}{how many components to model}

\item{init.method}{initialization method ('fuzzy' is new to initz2)}

\item{minprop}{minimum proportion of data in a component before fuzzing}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a list with values "pi", "mu", and "sd" of length ncomp
}\if{html}{\out{</div>}}
}
\description{
This was Zach Madaj's idea and it fixes the underlying issue in mixR, which
as of April 2024, remained in spite of previous attempts to patch it. Turns
out this is the least worst way to deal with a recurrent issue in 1D models.
}
\details{
nasty kludge:
}
\section{unlockBinding("initz", as.environment("package:mixR"))}{

}

\section{assign("initz", initz2, "package:mixR")}{
Hopefully we can eliminate the nasty kludge by contributing this to mixR.
}

\seealso{
ppclust::fpppcm

mixR::initz
}
