% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEapp.R
\name{iSEEapp}
\alias{iSEEapp}
\title{iSEEapp}
\usage{
iSEEapp(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment.}
}
\description{
A simple wrapper function to launch an iSEE web application for dxsmall.
iSEE is a Bioconductor package that you must install to run this function.
}
\examples{
if (require("iSEE")) {
  data(dxsmall)
  iSEEapp(dxsmall)
} 

}
\seealso{
SingleCellExperiment

dxsmall

iSEE
}
