% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanDiff.R
\name{meanDiff}
\alias{meanDiff}
\title{meanDiff}
\usage{
meanDiff(y, X)
}
\arguments{
\item{y}{Generic name for the outcome measures}

\item{X}{Generic name for the 2-level grouping variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Absolute value of the group 1 vs group 2 mean difference

}\if{html}{\out{</div>}}
}
\description{
meanDiff
}
