% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_beta.R
\name{parameters_beta}
\alias{parameters_beta}
\title{hook for beta/Gaussian equivalency tricks}
\usage{
parameters_beta(dat, ...)
}
\arguments{
\item{dat}{a vector of data, purportedly from a Beta distribution}

\item{...}{additional arguments to be passed on to MASS::fitdistr}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an object of class `fitdistr` (see MASS::fitdistr for more)
}\if{html}{\out{</div>}}
}
\description{
The canonical Beta distribution is expressed as
}
\details{
X ~ Beta(a, b)

where a > 0 is a shape parameter and b > 0 is another shape parameter.
\code{a} often represents the number of successes and \code{b} the number of failures
in a series of trials, whereupon the Beta represents the conjugate prior for
the binomial distribution where the success probability \code{p} is (a / (a + b)).
The MASS::fitdistr function requires start parameters for a Beta distribution
which we compute via the method of moments and then feed to MASS::fitdistr().
}
\examples{

set.seed(1234)
dat <- rbeta(n=1000, 17, 39)
parameters_beta(dat)

}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    MASS::fitdistr
}\if{html}{\out{</div>}}
}
