% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_beta.R
\name{params_beta}
\alias{params_beta}
\title{hook for distr6 beta/Gaussian equivalency tricks}
\usage{
params_beta(dat)
}
\arguments{
\item{dat}{a vector of data, purportedly from a univariate distribution}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     parameters for the closest beta-distributed approximation
}\if{html}{\out{</div>}}
}
\description{
hook for distr6 beta/Gaussian equivalency tricks
}
\examples{

set.seed(1234)
(params1 <- params_beta(rnorm(n=1e6)))

set.seed(1234)
# assume a standard beta VST in each direction
(params2 <- params_beta(bifurcatoR:::.expit(rnorm(n=1e6))))

}
