% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_weibull.R
\name{params_weibull}
\alias{params_weibull}
\title{hook for Weibull/Gaussian equivalency tricks}
\usage{
params_weibull(dat)
}
\arguments{
\item{dat}{a vector of data, purportedly from a univariate distribution}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     parameters for the closest Weibull-distributed approximation
}\if{html}{\out{</div>}}
}
\description{
hook for Weibull/Gaussian equivalency tricks
}
\examples{

set.seed(1234)
(params1 <- params_weibull(rnorm(n=1e6)))

set.seed(1234)
# assume a standard Weibull VST in each direction
(params2 <- params_weibull(log(pmax(0, rnorm(n=1e6)))))

}
