% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{permutation_tests}
\alias{permutation_tests}
\title{permutation_tests}
\usage{
permutation_tests(y, X, nperm, fxn_names, alpha)
}
\arguments{
\item{nperm}{The number of permutations to be performed}

\item{alpha}{The confidence level}

\item{temp_df}{Generic name for the data frame to be permuted columns are the group variable X and the value variable y}

\item{fxn}{The function name to call is \link{meanDiff} \link{madDiff} \link{giniDiff} \link{sdDiff}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a 3 level list containing: a p-value, the actual difference, and the 2.5\% - 97.5\% interval
}\if{html}{\out{</div>}}
}
\description{
permutation_tests
}
