% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixnorm.R
\name{rmixnorm}
\alias{rmixnorm}
\title{generate a random mixture of normals}
\usage{
rmixnorm(n, p1, p2, shift, sel, CI)
}
\arguments{
\item{n}{number of \link{something}}

\item{p1}{pr(event) in subgroup 1}

\item{p2}{pr(event) in subgroup 2}

\item{shift}{scaled distance between means}

\item{sel}{column name to select}

\item{CI}{generate a CI (?)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a data.frame
}\if{html}{\out{</div>}}
}
\description{
generate a random mixture of normals
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  let's see how dplyr verbs work in production :-O
}\if{html}{\out{</div>}}
}
\examples{
pps = rmixnorm(20, p1=0.25, p2=0.5, shift=15, sel='Bodyweight', CI=0)

}
