% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdDiff.R
\name{sdDiff}
\alias{sdDiff}
\title{sdDiff}
\usage{
sdDiff(y, X)
}
\arguments{
\item{y}{Generic name for the outcome measures}

\item{X}{Generic name for the 2-level grouping variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Absolute value of the difference between group 1 and group 2 standard deviation
}\if{html}{\out{</div>}}
}
\description{
sdDiff
}
