% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcatoR_Analysis.R
\name{bifurcatoR_Analysis}
\alias{bifurcatoR_Analysis}
\title{bifurcatoR_Analysis}
\usage{
bifurcatoR_Analysis(data, tests, nboot, alpha)
}
\arguments{
\item{data}{data frame with 2 columns, a two-level group column and numeric value column}

\item{tests}{names of tests to run}

\item{nboot}{number of bootstraps or permutations}

\item{alpha}{default significance level (0.05)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a data frame where each row corresponds to the results of test, p-values, test stats, and confidence intervals where possible
}\if{html}{\out{</div>}}
}
\description{
bifurcatoR_Analysis
}
