% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dxsmall}
\alias{dxsmall}
\title{dxsmall}
\format{
A \code{SummarizedExperiment} object
}
\source{
This data was created using inst/extdata/mRNA/dxsmall.reassemble.R
}
\usage{
data(dxsmall)
}
\description{
A SummarizedExperiment with covariates and mRNA counts from 501 AML cases.
}
\examples{

 data(dxsmall)
 show(dxsmall)

 rownames(dxsmall)
 names(colData(dxsmall))
 table(dxsmall$FusionGroup)
 table(dxsmall$AgeGroup)

 # if `mclust` is installed:
 if (require("mclust")) {

   fit1 <- Mclust(logcounts(dxsmall)["MECOM", dxsmall$FusionGroup == "MLL"])
   plot(fit1, what="density", xlab="log(MECOM read counts + 1) in KMT2Ar")

   fit2 <- Mclust(t(logcounts(dxsmall)[ c("MECOM", "PRDM16"), ]), G=1:3)
   plot(fit2, what="classification")
   plot(fit2, what="uncertainty")

 }


}
\keyword{datasets}
