% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow.R
\name{est_pow}
\alias{est_pow}
\title{est_pow}
\usage{
est_pow(n, p, alpha = 0.05, x, nsim = 20)
}
\arguments{
\item{n}{number of \link{somethings}}

\item{p}{pr(event)}

\item{alpha}{default significance level (0.05)}

\item{x}{I have no idea what this does}

\item{nsim}{number of simulations (20)}
}
\value{
\preformatted{   a power estimate
}
}
\description{
est_pow
}
