% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow.R
\name{est_pow}
\alias{est_pow}
\title{est_pow}
\usage{
est_pow(n, alpha, nsim, dist, params, tests)
}
\arguments{
\item{n}{number of \link{somethings}}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}

\item{dist}{generating distribution}

\item{params}{parameters for the generating distribution}

\item{tests}{names of tests to run}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a power estimate
}\if{html}{\out{</div>}}
}
\description{
est_pow
}
