% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pow_2samp.R
\name{est_pow_2samp}
\alias{est_pow_2samp}
\title{est_pow_2samp}
\usage{
est_pow_2samp(n1, n2, alpha, nsim, modes, dist, params, tests, nperm)
}
\arguments{
\item{n1}{number of \link{somethings} in group 1}

\item{n2}{number of \link{somethings} in group 2}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}

\item{modes}{modes of the two groups}

\item{dist}{generating distributions}

\item{params}{parameters for the generating distributions}

\item{tests}{names of tests to run between the two groups}

\item{nperm}{number of permutations to run}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a power estimate
}\if{html}{\out{</div>}}
}
\description{
est_pow_2samp
}
