% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_cs.R
\name{pow_cs}
\alias{pow_cs}
\title{calculate power several times}
\usage{
pow_cs(ns, p1, p2, shift, sel, CI, alpha = 0.05, nsim = 20)
}
\arguments{
\item{ns}{numbers of somethings}

\item{p1}{pr(event) in subgroup 1}

\item{p2}{pr(event) in subgroup 2}

\item{shift}{scaled distance between means}

\item{sel}{column to select}

\item{CI}{generate a CI (?)}

\item{alpha}{default significance level (0.05)}

\item{nsim}{number of simulations (20)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a data.frame 
}\if{html}{\out{</div>}}
}
\description{
calculate power several times
}
\examples{
pow_cs(ns=c(40,50), p1=.25, p2=.5, shift=15, sel='Bodyweight', CI=0)

}
